refactor logging and add setter
