--- config/params.ts
@@ -1,5 +1,0 @@
-export const params: string[] = []
-
-export function addParam(name: string) {
-  params.push(name)
-}
