--- config/settings.json
@@ -1,8 +1,0 @@
-{
-  "logging": {
-    "includeTimestamp": true,
-    "logFailure": true,
-    "logSuccess": true,
-    "logDataOnFailure": true
-  }
-}
