+++ functions/params.ts
@@ -1,0 +1,5 @@
+export const params: string[] = []
+
+export function addParam<DataType>(name: keyof DataType) {
+  params.push(name as string)
+}
